Fixed problem with type overloading

+++ tests/end_to_end_tests/src/integral_matrix_driver.cpp
@@ -9,11 +9,11 @@ int main (int argv, char** argc)
     size_t order = 0;
     std::cin >> order;
 
-    std::vector<int> data_vec (order * order);
+    std::vector<long long int> data_vec (order * order);
     for (auto& x : data_vec)
         std::cin >> x;
 
-    Matrix<int> M {order, order, data_vec.cbegin(), data_vec.cend()};
+    Matrix<long long int> M {order, order, data_vec.cbegin(), data_vec.cend()};
     std::cout << M.determinant() << std::endl;
 
     return 0;
